moved MainActivity into original package (prevent launcher icon disappear), fixed layout issues
