+++ app/src/main/kotlin/cz/covid19cz/erouska/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package cz.covid19cz.erouska.ui.ragnarok
+package cz.covid19cz.erouska.ui.main
 
 import android.content.Intent
 import android.net.Uri
@@ -7,11 +7,11 @@ import cz.covid19cz.erouska.R
 import cz.covid19cz.erouska.databinding.ActivityRagnarokBinding
 import cz.covid19cz.erouska.ui.base.BaseActivity
 import cz.covid19cz.erouska.ui.base.UrlEvent
+import cz.covid19cz.erouska.ui.ragnarok.RagnarokVM
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : BaseActivity<ActivityRagnarokBinding, RagnarokVM>(R.layout.activity_ragnarok, RagnarokVM::class) {
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
